move three methods from api-testing/response-handler to api-docs/specification-handler
- they are only used there
- now the OpenapiSpecificationHandler does not need a connection to the ResponseHandler
- this should make the split between those two easier
