+++ src/hackingBuddyGPT/utils/openapi/__init__.py
@@ -1,3 +1,1 @@
-from .openapi_converter import OpenAPISpecificationConverter
 from .openapi_parser import OpenAPISpecificationParser
-from .yaml_assistant import YamlFileAssistant
